requirements.txt
requests>=2.32
beautifulsoup4>=4.12
lxml>=5.2
playwright>=1.44
anthropic>=0.30
python-dotenv>=1.0
